consolidated imports and exports for screens and components into singular index.jsx files
