+++ frontend/src/components/index.jsx
@@ -1,0 +1,11 @@
+import LoginSignupTab from "./login_signup_tab";
+import Login from "./login";
+import SignUp from "./signup";
+import FoodTruckCard from "./food_truck_card";
+
+export {
+    LoginSignupTab,
+    Login,
+    SignUp,
+    FoodTruckCard,
+}
