+++ frontend/src/screens/index.jsx
@@ -1,0 +1,7 @@
+import SignIn from "./SignIn"
+import Home from "./Home"
+
+export {
+    SignIn,
+    Home
+}
